Fullscreen mode for media player in html5
The fullscreen mode in webkit is archived by hiding all the other
widgets besides the WebView widget.

We cannot simply reparent the WebView to a stand alone toplevel
GtkWindows. If we do that, it video will blackout or miss place
somewhere in the lower half of screen. It might has something to do with
XOverlay of the gstreamer which webkit used to play fullscreen media.

The "leaving-fullscreen" signal was not always emitted. When a user skim
to next feed item with hotkey(ctrl-n), that signal will not be emit.

To hack around the problem, the "load-status" of WebView is monitored
and the WebView is forced leaving fullscreen mode if a new page is
loaded into the WebView.

Another way to cope with it will be ignore all the liferea actions in
the fullscreen mode. But how to do that that?